Move functionality to framework
